fix: interaal/localstorage выделить в отдельный файл интерфейс Storager
